pyproject.toml
[project]
name = "tg-bot"
version = "0.2.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.15"
dependencies = [
    "aiogram (>=3.27.0,<4.0.0)",
    "python-dotenv (>=1.0.1)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
